requirements.txt
langgraph==0.2.55
langchain==0.3.12
langchain-core==0.3.25
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.3
streamlit==1.40.2
pandas==2.2.3
openai>=1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
ratelimit==2.2.1
aiohttp==3.9.1
setuptools>=65.0.0

